% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracy_score}
\alias{accuracy_score}
\title{Calculate Accuracy}
\usage{
accuracy_score(y_true, y_pred)
}
\arguments{
\item{y_true}{A vector of true labels}

\item{y_pred}{A vector of predicted labels}
}
\value{
A numeric value representing accuracy
}
\description{
Computes the classification accuracy given true and predicted labels.
}
