% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_models}
\alias{compare_models}
\title{Compare Models}
\usage{
compare_models(formula, data, k = 5)
}
\arguments{
\item{formula}{A formula specifying the model}

\item{data}{A data frame}

\item{k}{Number of folds (default = 5)}
}
\value{
A data frame comparing model performance
}
\description{
Compares logistic regression and SVM using k-fold CV.
}
