% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Generate Confusion Matrix}
\usage{
confusion_matrix(y_true, y_pred)
}
\arguments{
\item{y_true}{A vector of true labels (0/1)}

\item{y_pred}{A vector of predicted labels (0/1)}
}
\value{
A 2x2 matrix with TP, FP, FN, TN
}
\description{
Computes confusion matrix for binary classification.
}
