% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{f1_score}
\alias{f1_score}
\title{Calculate F1 Score}
\usage{
f1_score(y_true, y_pred)
}
\arguments{
\item{y_true}{A vector of true labels}

\item{y_pred}{A vector of predicted labels}
}
\value{
A numeric value representing F1 score
}
\description{
Computes F1 score for binary classification.
}
