% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{k_fold_cv}
\alias{k_fold_cv}
\title{K-Fold Cross Validation}
\usage{
k_fold_cv(formula, data, k = 5, model_type = "logistic")
}
\arguments{
\item{formula}{A formula specifying the model}

\item{data}{A data frame}

\item{k}{Number of folds (default = 5)}

\item{model_type}{"logistic" or "svm"}
}
\value{
A numeric value representing average accuracy
}
\description{
Performs k-fold cross validation for a given model.
}
