% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot Confusion Matrix}
\usage{
plot_confusion_matrix(y_true, y_pred)
}
\arguments{
\item{y_true}{A vector of true labels}

\item{y_pred}{A vector of predicted labels}
}
\value{
A ggplot object
}
\description{
Visualizes confusion matrix as a heatmap using ggplot2.
}
