% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{precision_score}
\alias{precision_score}
\title{Calculate Precision}
\usage{
precision_score(y_true, y_pred)
}
\arguments{
\item{y_true}{A vector of true labels (0/1 or factor)}

\item{y_pred}{A vector of predicted labels (0/1 or factor)}
}
\value{
A numeric value representing precision
}
\description{
Computes precision for binary classification.
}
