% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{recall_score}
\alias{recall_score}
\title{Calculate Recall}
\usage{
recall_score(y_true, y_pred)
}
\arguments{
\item{y_true}{A vector of true labels (0/1 or factor)}

\item{y_pred}{A vector of predicted labels (0/1 or factor)}
}
\value{
A numeric value representing recall
}
\description{
Computes recall for binary classification.
}
