% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_expanded}
\alias{titanic_expanded}
\title{Titanic Dataset}
\format{
A data frame with variables:
\describe{
  \item{Class}{Passenger class}
  \item{Sex}{Gender}
  \item{Age}{Age group}
  \item{Survived}{Survival status (0/1)}
}
}
\source{
R built-in Titanic dataset
}
\usage{
titanic_expanded
}
\description{
A processed version of the Titanic dataset for classification tasks.
}
\keyword{datasets}
