% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_logistic_model}
\alias{train_logistic_model}
\title{Train Logistic Regression Model}
\usage{
train_logistic_model(formula, data)
}
\arguments{
\item{formula}{A formula specifying the model}

\item{data}{A data frame}
}
\value{
A trained logistic regression model
}
\description{
Trains a logistic regression model using glm.
}
