% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_svm_model}
\alias{train_svm_model}
\title{Train SVM Model}
\usage{
train_svm_model(formula, data)
}
\arguments{
\item{formula}{A formula specifying the model}

\item{data}{A data frame}
}
\value{
A trained SVM model
}
\description{
Trains a support vector machine model.
}
